everything converted, working on optimizations and cleanup
