--- instrumentation/logback/logback-mdc-1.0/javaagent/src/addBaggageTest/groovy/io/opentelemetry/javaagent/instrumentation/logback/v1_0/LogbackWithBaggageTest.groovy
@@ -1,12 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.logback.v1_0
-
-import io.opentelemetry.instrumentation.logback.mdc.v1_0.AbstractLogbackWithBaggageTest
-import io.opentelemetry.instrumentation.test.AgentTestTrait
-
-class LogbackWithBaggageTest2 extends AbstractLogbackWithBaggageTest implements AgentTestTrait {
-}
