--- instrumentation/logback/logback-mdc-1.0/javaagent/src/test/groovy/io/opentelemetry/javaagent/instrumentation/logback/v1_0/LogbackTest.groovy
@@ -1,12 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.logback.v1_0
-
-import io.opentelemetry.instrumentation.logback.mdc.v1_0.AbstractLogbackTest
-import io.opentelemetry.instrumentation.test.AgentTestTrait
-
-class LogbackTest extends AbstractLogbackTest implements AgentTestTrait {
-}
