+++ instrumentation/logback/logback-mdc-1.0/library/src/addBaggageTest/java/io/opentelemetry/instrumentation/logback/mdc/v1_0/LogbackWithBaggageTest.java
@@ -5,4 +5,9 @@
 
 package io.opentelemetry.instrumentation.logback.mdc.v1_0;
 
-class LogbackWithBaggageTest extends AbstractLogbackWithBaggageTest {}
+class LogbackWithBaggageTest extends AbstractLogbackLibraryTest {
+  @Override
+  boolean expectBaggage() {
+    return true;
+  }
+}
