+++ instrumentation/logback/logback-mdc-1.0/library/src/test/java/io/opentelemetry/instrumentation/logback/mdc/v1_0/LogbackTest.java
@@ -5,4 +5,4 @@
 
 package io.opentelemetry.instrumentation.logback.mdc.v1_0;
 
-class LogbackTest extends AbstractLogbackTest {}
+class LogbackTest extends AbstractLogbackLibraryTest {}
