--- instrumentation/logback/logback-mdc-1.0/testing/src/main/java/io/opentelemetry/instrumentation/logback/mdc/v1_0/AbstractLogbackWithBaggageTest.java
@@ -1,13 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.logback.mdc.v1_0;
-
-public abstract class AbstractLogbackWithBaggageTest extends AbstractLogbackTest {
-  @Override
-  boolean expectBaggage() {
-    return true;
-  }
-}
